+++ agora/agora/utils.py
@@ -1,28 +1,22 @@
 import os
 import json
 import re
-from bs4 import BeautifulSoup, Comment
+from collections import defaultdict
 from HTMLParser import HTMLParseError
+from bs4 import BeautifulSoup, Comment
 from ckeditor_uploader.fields import RichTextUploadingField
 
-from collections import defaultdict
 from argo_ams_library import ArgoMessagingService, AmsMessage, AmsException
 
 from django.conf import settings
-
-from collections import defaultdict
-
-from django.conf import settings
-from agora.permissions import get_rules
 import accounts.models
-
+from agora.permissions import get_rules
 from agora.settings import (
-     AMS_TOKEN,
-     AMS_ENDPOINT,
-     AMS_PROJECT,
-     AMS_TOPIC,
+    AMS_TOKEN,
+    AMS_ENDPOINT,
+    AMS_PROJECT,
+    AMS_TOPIC,
 )
-
 
 
 _root_url = None
@@ -92,9 +86,23 @@ _root_url = deploy_config[':root_url']
 def get_root_url():
     return _root_url
 
-def publishMessage(instance):
-    ams = ArgoMessagingService(endpoint=AMS_ENDPOINT, project=AMS_PROJECT, token=AMS_TOKEN)
-    endpoint =  '{0}{1}'.format(get_root_url(), '/api/v2/ext-services')
+
+def publish_message(service, action):
+    """
+    Send a message using argo messaging service when an action upon a service
+    takes place.
+    """
+    service_id = str(service.get('id'))
+    service_name = service.get('name')
+    service_data = service.get('data', {})
+    ams = ArgoMessagingService(endpoint=AMS_ENDPOINT,
+                               project=AMS_PROJECT,
+                               token=AMS_TOKEN)
+    endpoint = '{0}/api/v2/ext-services/{1}'.format(get_root_url(), service_id)
+    # The value of the data property must be unicode in order to be
+    # encoded in base64 format in AmsMessage
+    data = json.dumps(service_data)
+
     try:
         if not ams.has_topic(AMS_TOPIC):
             ams.create_topic(AMS_TOPIC)
@@ -102,12 +110,19 @@ def publishMessage(instance):
         print e
         raise SystemExit(1)
 
-    msg = AmsMessage(data=endpoint, attributes={'bar1': 'baz1'}).dict()
+    msg = AmsMessage(data=data,
+                     attributes={
+                         "method": action,
+                         "service_id": service_id,
+                         "service_name": service_name,
+                         "endpoint": endpoint
+                     }).dict()
     try:
         ret = ams.publish(AMS_TOPIC, msg)
         print ret
     except AmsException as e:
         print e
+
 
 def clean_html_fields(instance):
     class_fields = instance.__class__._meta.get_fields()
@@ -161,7 +176,7 @@ def safe_html(html):
             tag.attrs = []
 
     # scripts can be executed from comments in some cases
-    comments = soup.findAll(text=lambda text:isinstance(text, Comment))
+    comments = soup.findAll(text=lambda text: isinstance(text, Comment))
     for comment in comments:
         comment.extract()